file_mapper add option to map an anonymous page after the file
This can be useful if you want to make sure your file is null terminated.
For example, if you're running over the contents of a text file,
this will allow you to use strchrnul without worrying about a segfault if
the file isn't null terminated and its size happens to fall on a page
boundary (mmap will round the map's size to the next page, so if the file
size isn't a multiple of a page boundary, mmap will automatically
null-terminate it).

Change-Id: I39673b4874f90387e7b4651675d4cf6c7f36d673