fix build in debian 12
Since ab0cd5f (catch SIGTERM via signalfd, 2014-03-10) a static stack
is provided for signal handlers using a system defined variable.

Linux has changed the definition of it to use a `sysconf()` call
instead as a way to allow variable stacks and therefore that would
break the build in recent systems.

Define a RIBS specific variable with equivalent value and use that
instead in cases where glibc does that redefinition or if it is not
defined at all (ex: alpine or voidlinux with musl).

While at it, do some minor cleanup in an unrelated header.